Fix zsh plugin loading
fzf-tab did not want to load when specified in the plugins array. In
fact it broke the loading of all other plugins as well.
I suspect it has something to do with the plugin file and the directory
it is in both being symlinks in my setup.

Loading fzf-tab separately for now instead.