Update Dockerfile
Install neovim from nightly binary instead of apt to get the latest
version, because stable is too old for some of my plugins.

Install python because it's a dependency for some vim plugins.

Configured zsh as the shell so it starts automatically when testing the
image.

Combined user creation commands to one layer.